feat: move SeedDta schema and update imports
