+++ fastapi_backend/db/seeds/seed.py
@@ -1,5 +1,5 @@
 from fastapi_backend.db.db_manager import  DBManager
-from fastapi_backend.schemas.seed_models import SeedData
+from fastapi_backend.db.schemas.seed_models import SeedData
 
 async def seed(
         db_manager: DBManager,
